revert apollo logger plugin as per review comments
