+++ abods-api/src/lib/datadog.ts
@@ -7,13 +7,11 @@ import {
   GraphQLRequestExecutionListener,
   GraphQLRequestListener,
   GraphQLRequestListenerDidResolveField,
-  GraphQLServerListener,
 } from "@apollo/server";
 import { RequestContext } from "../types/extra";
 import { sendDistributionMetric } from "datadog-lambda-js";
 import { PrismaClientInitializationError } from "@prisma/client/runtime/library";
 import { GraphQLError } from "graphql";
-import logger from "../logger.js";
 
 export const sendErrorMetric = (error: unknown) => {
   if (
@@ -40,47 +38,22 @@ export const sendErrorMetric = (error: unknown) => {
   );
 };
 
-const serverListener: GraphQLServerListener = {
-  drainServer: () => {
-    logger.info("Draining server");
-    return Promise.resolve();
-  },
-  serverWillStop: () => {
-    logger.info("Stopping server");
-    return Promise.resolve();
-  },
-  schemaDidLoadOrUpdate: () => logger.info("Schema loaded"),
-};
-
 const datadogMetricsPlugin: ApolloServerPlugin = {
   invalidRequestWasReceived({ error }) {
     sendErrorMetric(error);
-    logger.error({ error }, "Received invalid request");
     return Promise.resolve();
   },
   startupDidFail({ error }) {
     sendErrorMetric(error);
-    logger.error({ error }, "Startup failed");
     return Promise.resolve();
   },
   contextCreationDidFail({ error }) {
-    logger.error({ error }, "Failed to create context");
     sendErrorMetric(error);
     return Promise.resolve();
   },
-  unexpectedErrorProcessingRequest({
-    error,
-    requestContext: {
-      request: { query, operationName },
-    },
-  }) {
+  unexpectedErrorProcessingRequest({ error, requestContext: _ }) {
     sendErrorMetric(error);
-    logger.error({ error, query, operationName }, "Failed to process request");
     return Promise.resolve();
-  },
-  serverWillStart() {
-    logger.info("Started server");
-    return Promise.resolve(serverListener);
   },
 
   requestDidStart: async (
@@ -89,27 +62,16 @@ const datadogMetricsPlugin: ApolloServerPlugin = {
     const startTime = Date.now();
 
     return {
-      didEncounterErrors: ({ errors, request: { query, operationName } }) => {
+      didEncounterErrors: ({ errors, request: _ }) => {
         for (const error of errors) {
-          logger.error(
-            { error, query, operationName },
-            "Encountered error when processing request",
-          );
           sendErrorMetric(error);
         }
         return Promise.resolve();
       },
 
-      didEncounterSubsequentErrors: ({
-        errors,
-        request: { query, operationName },
-      }) => {
+      didEncounterSubsequentErrors: ({ errors, request: _ }) => {
         if (!errors) return Promise.resolve();
         for (const error of errors) {
-          logger.error(
-            { error, query, operationName },
-            "Encountered error when processing request",
-          );
           sendErrorMetric(error);
         }
         return Promise.resolve();
